funcionalidad de actualizar productos movida a products
